+++ doc/user/Changelog.md
@@ -94,6 +94,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Markdown prettier formatting [#7815](https://github.com/opentripplanner/OpenTripPlanner/pull/7815)
 - Report invalid line ref as DataImportIssue in NeTEx TripMapper [#7790](https://github.com/opentripplanner/OpenTripPlanner/pull/7790)
 - Fix ArrayIndexOutOfBoundsException in pass-through via chaining [#7747](https://github.com/opentripplanner/OpenTripPlanner/pull/7747)
+- Snapshot framework [#7689](https://github.com/opentripplanner/OpenTripPlanner/pull/7689)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.9.0 (2026-03-18)